xinSLM_v05_distillation/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
tokenizers>=0.14.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.14.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
huggingface-hub>=0.17.0
peft>=0.6.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.21.0
pyyaml>=6.0
jsonlines>=3.1.0
psutil>=5.9.0
gpustat>=1.1.0
flash-attn>=2.3.0; sys_platform != "darwin"
optimum>=1.13.0
auto-gptq>=0.4.2
safetensors>=0.4.0
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.1
nltk>=3.8.1
torchmetrics>=1.2.0